# frontend/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Physionexs API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/clinic/queue": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "Get Queue",
        "operationId": "get_queue_clinic_queue_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "branch_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Branch Id"
            }
          },
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Register Walk In",
        "operationId": "register_walk_in_clinic_queue_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalkInIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clinic/queue/call-next": {
      "post": {
        "tags": [
          "queue"
        ],
        "summary": "Call Next",
        "operationId": "call_next_clinic_queue_call_next_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "branch_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Branch Id"
            }
          },
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clinic/queue/{token_id}": {
      "patch": {
        "tags": [
          "queue"
        ],
        "summary": "Update Token",
        "operationId": "update_token_clinic_queue__token_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Token Id"
            }
          },
          {
            "name": "X-Clinic-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Clinic-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueTokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/tokens": {
      "get": {
        "tags": [
          "queue"
        ],
        "summary": "My Tokens",
        "description": "The patient's walk-in tokens that are still live today, with position and estimated wait.",
        "operationId": "my_tokens_me_tokens_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MyTokenOut"
                  },
                  "type": "array",
                  "title": "Response My Tokens Me Tokens Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyTokenOut": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "$ref": "#/components/schemas/TokenStatus"
          },
          "clinic_name": {
            "type": "string",
            "title": "Clinic Name"
          },
          "branch_name": {
            "type": "string",
            "title": "Branch Name"
          },
          "ahead": {
            "type": "integer",
            "title": "Ahead"
          },
          "est_wait_minutes": {
            "type": "integer",
            "title": "Est Wait Minutes"
          },
          "now_serving": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Now Serving"
          }
        },
        "type": "object",
        "required": [
          "label",
          "status",
          "clinic_name",
          "branch_name",
          "ahead",
          "est_wait_minutes",
          "now_serving"
        ],
        "title": "MyTokenOut"
      },
...
      "QueueOut": {
        "properties": {
          "branch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Branch Id"
          },
          "service_date": {
            "type": "string",
            "format": "date",
            "title": "Service Date"
          },
          "now_serving": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueueTokenOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "waiting": {
            "items": {
              "$ref": "#/components/schemas/QueueTokenOut"
            },
            "type": "array",
            "title": "Waiting"
          },
          "done": {
            "items": {
              "$ref": "#/components/schemas/QueueTokenOut"
            },
            "type": "array",
            "title": "Done"
          },
          "avg_wait_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Wait Minutes"
          }
        },
        "type": "object",
        "required": [
          "branch_id",
          "service_date",
          "now_serving",
          "waiting",
          "done",
          "avg_wait_minutes"
        ],
        "title": "QueueOut"
      },
      "QueueTokenOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "status": {
            "$ref": "#/components/schemas/TokenStatus"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "clinic_patient_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clinic Patient Id"
          },
          "patient_name": {
            "type": "string",
            "title": "Patient Name"
          },
          "patient_age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Age"
          },
          "patient_sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sex"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "physio_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physio Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "called_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Called At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "number",
          "status",
          "patient_id",
          "clinic_patient_id",
          "patient_name",
          "patient_age",
          "patient_sex",
          "reason",
          "physio_name",
          "created_at",
          "called_at"
        ],
        "title": "QueueTokenOut"
      },
...
      "Sex": {
        "type": "string",
        "enum": [
          "M",
          "F",
          "O"
        ],
        "title": "Sex"
      },
...
      "TokenStatus": {
        "type": "string",
        "enum": [
          "waiting",
          "serving",
          "done",
          "skipped"
        ],
        "title": "TokenStatus"
      },
      "TokenStatusIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TokenStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TokenStatusIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WalkInIn": {
        "properties": {
          "branch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Branch Id"
          },
          "patient_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patient Id"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sex"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "physio_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physio Id"
          }
        },
        "type": "object",
        "required": [
          "branch_id"
        ],
        "title": "WalkInIn"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}